Ensure stack maps are 4 byte aligned.
With the recent move to gcc 4.9, we are hitting alignment
SIGBUS on ARM. The reason is that gcc will optimize two consecutive
32bits loads into one 64bits load, and the instruction (ldrd)
will fail if the data is not aligned.

Also removed the emission of mapping table when a method is optimized.
The information can be found in the StackMap itself.

Change-Id: Icf79406c18a3f4db3c05d52fc2c0dd2e35bf0f8f

Backported to Android 5.0 Lollipop by arter97.
Original commit : https://android.googlesource.com/platform/art/+/376b2bbf7c39108223a7a01568a7b4b04d84eeac

Signed-off-by: arter97 <qkrwngud825@gmail.com>