ART: patchoat gets gzip support for compressed odex files
The patchoat is capable to inflate a compressed odex file
into a temporary file. Then the temporary file is patched
as if it is specified with --input-oat-fd or --input-oat-file.

The new options are the following:
  --input-oat-gz-fd
  --input-oat-gz-file
  --swap-fd
  --swap-file

Could not make in-memory decompression and patching because the ElfFile
class maps Elf file using its file descriptor. So a temporary file is
used as it is done by dex2oat.

Both gzip and zip are supported by zlib library and they have the same
compression algorithm. Gzip is chosen because zip needs a zip entry name,
which is unnecessary for single-file compression.

Change-Id: I54dbfcdc58bc0f57f7eccc3929c09295c597794b
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>

Conflicts:

	patchoat/Android.mk